clean the process of init

+++ install.sh
@@ -1,56 +1,24 @@
 #! /bin/bash
-if [ -d ~/dotfiles ]; then
-    if [ ! -e ~/.vimrc ]; then
-        ln -s ~/dotfiles/.vimrc ~/.vimrc
-    elif [ ! -e ~/.vim ]; then
-        ln -s ~/dotfiles/.vim ~/.vim
-        ln -s ~/dotfiles/.vim/.vimrc.basic ~/.vim
-        ln -s ~/dotfiles/.vim/.vimrc.bundle ~/.vim
-        ln -s ~/dotfiles/.vim/.vimrc.color ~/.vim
-        ln -s ~/dotfiles/.vim/.vimrc.indent ~/.vim
-        ln -s ~/dotfiles/.vim/.vimrc.key ~/.vim
-        ln -s ~/dotfiles/.vim/.vimrc.path ~/.vim
-        ln -s ~/dotfiles/.vim/.vimrc.plugin ~/.vim
-        ln -s ~/dotfiles/.vim/.vimrc.theme ~/.vim
-        ln -s ~/dotfiles/.vim/colors/solarized.vim ~/.vim/colors
-        ln -s ~/dotfiles/.vim/bundle ~/.vim
-    elif [ ! -e ~/.config ]; then
-        ln -s ~/dotfiles/.config ~/.config
-    elif [ ! -e ~/.tmux.conf ]; then
-        ln -s ~/dotfiles/.tmux.conf ~/.tmux.conf
-    elif [ ! -e ~/.viminfo ]; then
-        ln -s ~/dotfiles/.viminfo ~/.viminfo
-    elif [ ! -e ~/.vimperator ]; then
-        ln -s ~/dotfiles/.vimperator ~/.vimperator
-    elif [ ! -e ~/.vimperatorrc ]; then
-        ln -s ~/dotfiles/.vimperatorrc ~/.vimperatorrc
-    elif [ ! -e ~/.zshrc ]; then
-        ln -s ~/dotfiles/.zshrc ~/.zshrc
-    fi
-else
-    mkdir ~/dotfiles 
+[ -e ~/.vimrc ] || ln -s ~/dotfiles/.vimrc ~/.vimrc
+if [ ! -e ~/.vim ]; then
+    ln -s ~/dotfiles/.vim ~
+    ln -s ~/dotfiles/.vim/colors/solarized.vim ~/.vim/colors
+    ln -s ~/dotfiles/.vim/bundle ~/.vim
+    files=( basic bundle color indent key path plugin theme )
+    for file in ${files[@]}; do
+        ln -s ~/dotfiles/.vim/.vimrc.$file ~/.vim/.vimrc.$file
+    done
 fi
+[ -e ~/.config ]       || ln -s ~/dotfiles/.config ~/.config
+[ -e ~/.tmux.conf ]    || ln -s ~/dotfiles/.tmux.conf ~/.tmux.conf
+[ -e ~/.viminfo ]      || ln -s ~/dotfiles/.viminfo ~/.viminfo
+[ -e ~/.vimperator ]   || ln -s ~/dotfiles/.vimperator ~/.vimperator
+[ -e ~/.vimperatorrc ] || ln -s ~/dotfiles/.vimperatorrc ~/.vimperatorrc
+[ -e ~/.zshrc ]        || ln -s ~/dotfiles/.zshrc ~/.zshrc
 
-if [ -e /usr/local/bin/brew ]; then
-    if [ ! -e /usr/local/bin/htop ]; then
-        brew install htop
-    elif [ ! -e /usr/local/bin/rbenv ]; then
-        brew install rbenv
-    elif [ ! -e /usr/local/bin/fish ]; then
-        brew install fish
-    elif [ ! -e /usr/local/bin/tig ]; then
-        brew install tig
-    elif [ ! -e /usr/local/bin/git ]; then
-        brew install git
-    elif [ ! -e /usr/local/bin/vim ]; then
-        brew install vim
-        brew install tidy-html5
-    elif [ ! -e /usr/local/bin/nodenv ]; then
-        brew install nodenv
-    elif [ ! -e /usr/local/bin/tmux ]; then
-        brew install tmux
-        brew install reattach-to-user-namespace
-    fi
+if [ ! $(which brew) ]; then
+    brew tap homebrew/brewdler
+    brew brewdle
 else
     /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 fi
